requirements.txt
aiohttp==3.8.3
beautifulsoup4==4.12.2
fastapi==0.111.0
numpy==1.24.3
openai==0.28.0
pika==1.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.29.0
uvicorn==0.30.0
pgvector==0.3.0
aiofiles==22.1.0
scikit-learn==1.2.2
